requirements.txt
boto3==1.4.4
botocore==1.5.76
certifi==2017.4.17
chardet==3.0.4
docutils==0.13.1
idna==2.5
jmespath==0.9.3
python-dateutil==2.6.0
PyYAML==3.12
requests==2.18.1
s3transfer==0.1.10
six==1.10.0
urllib3==1.21.1
